Split APIHandler to api.py
